[NAYB-37] refactor : dto 파일 이동
