+++ src/main/java/com/prgrms/nabmart/domain/event/service/request/RegisterEventCommand.java
@@ -1,4 +1,4 @@
-package com.prgrms.nabmart.domain.event.dto;
+package com.prgrms.nabmart.domain.event.service.request;
 
 public record RegisterEventCommand(String title, String description) {
